[ REFACTOR ] Interview Model 네이밍 변경
