+++ app/src/main/java/com/lottomate/lottomate/presentation/screen/interview/InterviewScreen.kt
@@ -54,8 +54,8 @@ import com.lottomate.lottomate.presentation.component.LottoMateIconButton
 import com.lottomate.lottomate.presentation.component.LottoMateText
 import com.lottomate.lottomate.presentation.component.LottoMateTopAppBar
 import com.lottomate.lottomate.presentation.res.Dimens
-import com.lottomate.lottomate.presentation.screen.interview.model.Interview
-import com.lottomate.lottomate.presentation.screen.interview.model.InterviewMockData
+import com.lottomate.lottomate.presentation.screen.interview.model.InterviewQnA
+import com.lottomate.lottomate.presentation.screen.interview.model.InterviewUIModel
 import com.lottomate.lottomate.presentation.ui.LottoMateBlack
 import com.lottomate.lottomate.presentation.ui.LottoMateGray100
 import com.lottomate.lottomate.presentation.ui.LottoMateGray120
